sari-store/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
Gunicorn==21.2.0
python-dotenv==1.0.0
